Move MainScene class to separated file
